Add files via upload
The fire class is a basic C++ class for creating an animation, in this case a simple fire simulation. The class is written to run a single red LED on a PWM pin, its brightness rising and falling over time with random bright flashes. The demonstration sketch runs two LEDS to demonstrate object independence which is critical to effective animations. For more information, see the blog post http://thenscaler.com/?p=1186